y20/src/buri/aoc/y20/d23/PuzzleTest.java
package buri.aoc.y20.d23;

import static org.junit.Assert.*;

import org.junit.Test;

import buri.aoc.BaseTest;
import buri.aoc.Part;

/**
 * @author Brian Uri!
 */
public class PuzzleTest extends BaseTest {

	@Test
	public void testPart1Examples() {
		assertEquals("67384529", Puzzle.getResult(Part.ONE, Puzzle.getInput(1).get(0)));
	}

	/**
	 * Solves the Part 1 puzzle.
	 */
	@Test
	public void testPart1Puzzle() {
		String result = Puzzle.getResult(Part.ONE, Puzzle.getInput(0).get(0));
		toConsole(result);
		assertEquals("45798623", result);
	}

	@Test
	public void testPart2Examples() {
		assertEquals("149245887792", Puzzle.getResult(Part.TWO, Puzzle.getInput(1).get(0)));
	}

	/**
	 * Solves the Part 2 puzzle.
	 */
	@Test
	public void testPart2Puzzle() {
		String result = Puzzle.getResult(Part.TWO, Puzzle.getInput(0).get(0));
		toConsole(result);
		assertEquals("235551949822", result);
	}
}
